Load yaml files with orphan refs
Specmatic eagerly loads the types of all refs. When it encounters a ref containing a name that is not defined as a component, it throws an exception.

The clean fix is to avoid loading all types eagerly. But this is going to take more time than I have.

For now I am simply returning an empty object. As an unfortunate result, it's possible for an empty object to match such a ref, while it should actually throw an exception. This should be fairly rare, so I'm moving ahead with this for now. But this needs to be fixed.