Show API in a match Failure error report
When debugging error reports, show anything that will make it easy to pinpoint which contract the payload failed to match. Specmatic already shows the contract file containing the partially matched API. Additionally, post the method, path and status so that we can narrow down the search for the payload within the contract.