Display API summary in junit test report
The legacy jnit xml report feature of JUnit5 uses the method name as the test name. At this time, there is no way to customise it. When the tests are generated dynamically, as they are with contract tests, customise the names by replacing the method-based name in the legacy xml report with the actual one, after the report has been generated.