Fix dictionary of array of strings
Specmatic parses expressions from right to left. Given (string*), it
would first pickup the *, identify this expression as an array of
something, and then find the type of that something to the left of the
star. Given this logic, (dictionary string string*) gets parsed as an
array of dictionaries of strings, whereas what we meant was a dictionary
of string arrays. If the expression is a dictionary, break it up first
into key and value types before parsing, rather than parsing the whole
expression first.