Fix to ephemeral contract repo checkout
Currently, cloning a repo is attempted first using jgit, and if that
doesn't work, Specmatic falls back to the git command.

JGit leaves the newly created directory in place, along with the .git
directory if the clone failed. This makes it look superficially as if
there is a repo in the new directory, since a .git directory exists
within it. The git command now does not want to override this and fails
 to checkout the repo on top of it.

Fairly sure this wasn't always the case, but this is certainly how it
works now.

Reset the directory in case of jgit checkout failure, so that the git
command sees an empty target directory.