First cut of dependency graph
Register the azure collection of the project in specmatic.json, and a
personal access token in a file in the home directory. Specmatic will
use this info to build a dependency graph from information available in
Azure.

The command "specmatic graph consumers" will list out the repositories
containing consumers that have stubbed out all contracts registered for
test in the project's specmatic.json file.

Currently this command supports Azure only.