Show stub errors for status-matched endpoints
If an endpoint in OpenAPI has declared multiple response statuses, it
translates to multiple scenarios. When a request does not match that
endpoint, we get match failures on all the scenarios.

Instead, since the request is fundamentally for a single OpenAPI
endpoint, show only the error from the scenario with the status matching
 the stub.