Custom implicit stub base as method
The val version was caching the content of environment variable and
property, which are both unset when the tests are run. When these are
changed in the tests, they are not picked up. The function however runs
every time it is called, and picks up any changes. We should eventually
come up with a better way of passing config than relying on this
function to fetch the current values.