Describe "" as "empty string"
Error message display the expected and actual value. Often the expected
value turns out to be empty for some reason, and resulted in an error
like "Contract expected json object but request contained", because an
empty string has no value and hence is invisible if it's value is
displayed. We now show a custom error message "an empty string", so the
error now reads like this: "Contract expected json object but request
contained an empty string"