Buffer first 3 events, fix flaky event generation
1. When an event is sent, at least the way we are using the Channel
object now, one consumer reads it, and not the others. This was
resulting in dropped events initially after closing one connection and
opening another. Fixed this by opening a new channel for every
connection, derive a broadcast channel from it, and store it in a list.
When a new SSE event comes in, iterate over the broadcast channel list
and send it to each channel.
2. When a channel is opened, it times out if the initial response is not
 sent. So we now with an empty response, and that holds the channel
 open.
3. When a connection is opened, it might be useful to send the an
initial buffer of events. Added a buffer of 3 events, but may alter this
 feature per need.