Add or update the SSE buffer
When a connection is made to an SSE url, an initial dump with the
current state of the system may be useful, and we now have a provision
for managing a buffer of events to send as an initial dump.

There is a new parameter now named bufferIndex.

* If bufferIndex is not sent, the event is sent to all subscribers.
* If bufferIndex is set to -1, the event is appended to the event
buffer, and not sent out to subscribers. New subscribers will be sent
the contents of the buffer.
* If the bufferIndex is >= 0, the buffer will be updated with this event
 at that index. This will be useful if the initial state sent to new
 subscribers needs to be updated due to any event sent to current
 subscribers.