Use the type system to signal whether a stub was found
We need to check if a stub match was found so that we know whether to
track it for coverage. The code is currently structured to return
responses. But with this change we are returning a stub result object
indicating success or failure, and we've moved the logging behaviour
into that object.