Used "first" instead of "single" in AnyPattern.listOf
listOf is intended to wrap the provided values in a list based on the
type of values. For example, a string would be wrapped in a
JSONArrayValue object. An XMLNode object would be wrapped in another
XMLNode object.

We can safely assume that all the types in an AnyPattern at any given
point due to the way that the object is constructed would consist of
values of the same type (all the values would be either JSON or XML).

Therefore, instead of using single (which can break given an enum type
 with multiple options), we use first to identify the element in the
 list of values that shall provide a wrapping list object.

This commit also includes a refactored version of the new tests.