Base64StringPattern.parse should not encode a non-base64 value
The intent of this method is only to convert a string to the right
data-type. This usually involves an implicit format check. For example,
a value like "pqr" cannot be parsed to a decimal number, and an
exception would result.

When the target type is a base64 string (which is actually a format), do
the necessary validation without changing the value being parsed.