When loading example as stub, use a fake the auth param
If a security scheme is provided, and an is available and used as a
stub, there's no way to provide an example of a security parameter. As a
result, at the moment the example is rejected at startup with an error.

Instead, use a dummy example in the stub, and let the stub match
requests that have the security parameter (without validating its
value)