fix(http-client): fix issue with leaving too many open connections
When running a close loop test, too many connections remain in
`TIME_WAIT` state. This is a side-effect of how tcp works. TCP
connections go into this state, in anticipation of any delayed packets
still on the network after the connection is closed.

Earlier implementation created a new httpclient (and connection) for
each HTTP request/response cycle, thereby making the situation worse.

Using a longer lived http-client tends to reuse the same TCP connection
for more than one HTTP call and significantly reduces (near zero) the
number of connections in `TIME_WAIT` state.

`HttpClient` is renamed to `Legacy` and the new `HttpClient` expects
the consumer to call `close` or use with `.use`